# src/main/resources/openapi/api.yaml
openapi: 3.0.3
info:
  title: Car Rental API
  description: Backend API for the car rental application.
  version: 1.0.0
servers:
  - url: /
paths:
...
  /api/users:
    get:
      tags:
        - Users
      operationId: listUsers
      summary: List all user accounts, optionally searching by name or email. Restricted to admin accounts.
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: search
          in: query
          description: Case-insensitive partial match against first name, last name, or email.
          schema:
            type: string
      responses:
        "200":
          description: A page of user accounts.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPage"
        "401":
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authenticated but not an admin account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
        - Users
      operationId: createUser
      summary: Create a new user account. Admins may create any role; dealership accounts may only create USER or DEALERSHIP accounts.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfileResponse"
        "400":
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authenticated but not an admin or dealership account, or a dealership account attempting to create an admin account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: An account already exists for this email.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/users/me:
    get:
      tags:
        - Users
      operationId: getCurrentUser
      summary: Return the profile of the currently authenticated user.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The current user's profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfileResponse"
        "401":
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/users/me/password:
    put:
      tags:
        - Users
      operationId: changePassword
      summary: Change the current user's own password.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "204":
          description: Password changed.
        "400":
          description: Validation error, or the current password is incorrect.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/reservations:
    post:
      tags:
        - Reservations
      operationId: createReservation
      summary: >
        Create a reservation for a car. Admins may reserve any car; dealership or user accounts
        may only reserve a car belonging to their own company.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReservationRequest"
      responses:
        "201":
          description: Reservation created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationResponse"
        "400":
          description: Validation error, or endTime is not after startTime.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: This car does not belong to the caller's company.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No car exists with this id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: >
            This car already has a reservation that overlaps, or is within 2 hours of, the
            requested time window.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/reservations/{id}:
    get:
      tags:
        - Reservations
      operationId: getReservationById
      summary: >
        Get the details of a reservation. Admins may view any reservation; dealership or user
        accounts may only view a reservation for a car belonging to their own company.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The reservation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationResponse"
        "401":
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: This reservation's car does not belong to the caller's company.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No reservation exists with this id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags:
        - Reservations
      operationId: updateReservation
      summary: >
        Update a reservation. Admins may update any reservation; dealership or user accounts may
        only update a reservation for a car belonging to their own company. The full reservation
        (including carId) is replaced, and moving a reservation to a different car requires access
        to that car too.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReservationRequest"
      responses:
        "200":
          description: Reservation updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationResponse"
        "400":
          description: Validation error, or endTime is not after startTime.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: >
            This reservation's car, or the target car, does not belong to the caller's company.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No reservation or no car exists with the given id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: >
            The target car already has a reservation that overlaps, or is within 2 hours of, the
            requested time window.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - Reservations
      operationId: deleteReservation
      summary: >
        Delete a reservation. Admins may delete any reservation; dealership or user accounts may
        only delete a reservation for a car belonging to their own company.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Reservation deleted.
        "401":
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: This reservation's car does not belong to the caller's company.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No reservation exists with this id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/cars/{carId}/reservations:
    get:
      tags:
        - Reservations
      operationId: listReservationsForCar
      summary: >
        List all reservations for a car. Admins may view any car's reservations; dealership or
        user accounts may only view reservations for a car belonging to their own company.
      security:
        - bearerAuth: []
      parameters:
        - name: carId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: A page of reservations for the car.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationPage"
        "401":
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: This car does not belong to the caller's company.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No car exists with this id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/companies/{companyId}/reservations:
    get:
      tags:
        - Reservations
      operationId: listReservationsForCompany
      summary: >
        List all reservations for a company's cars. Admins may view any company's reservations;
        dealership or user accounts may only view reservations for their own company.
      security:
        - bearerAuth: []
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: A page of reservations across the company's cars.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationPage"
        "401":
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: This company is not the caller's own.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No company exists with this id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserRole:
      type: string
      description: All roles a user account can have.
      enum:
        - USER
        - DEALERSHIP
        - ADMIN
...
    ChangePasswordRequest:
      type: object
      required:
        - currentPassword
        - newPassword
        - newPasswordConfirmation
      properties:
        currentPassword:
          type: string
          format: password
        newPassword:
          type: string
          format: password
          minLength: 8
        newPasswordConfirmation:
          type: string
          format: password
          minLength: 8
    CreateUserRequest:
      type: object
      required:
        - email
        - password
        - firstName
        - lastName
        - role
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        role:
          $ref: "#/components/schemas/UserRole"
        companyId:
          type: string
          description: >
            Company to attach this user to. Required when an admin creates a USER or DEALERSHIP account
            (rejected as missing otherwise). Ignored when a dealership account creates the user: the new
            user is always attached to the caller's own company instead. Not used when creating an ADMIN account.
...
    UserProfileResponse:
      type: object
      required:
        - id
        - email
        - firstName
        - lastName
        - role
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        companyId:
          type: string
    UserPage:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/UserProfileResponse"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
    ErrorResponse:
      type: object
      required:
        - status
        - message
        - timestamp
      properties:
        status:
          type: integer
        message:
          type: string
        timestamp:
          type: string
          format: date-time
...
    CreateReservationRequest:
      type: object
      required:
        - carId
        - startTime
        - endTime
      properties:
        carId:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
    ReservationResponse:
      type: object
      required:
        - id
        - carId
        - startTime
        - endTime
      properties:
        id:
          type: string
        carId:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
    ReservationPage:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/ReservationResponse"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
...